fix: add facility org setup and fix activity definition slug
- Add facilityOrg.setup.ts to ensure care-fac-admin and care-nurse
  are in the Administration org, and Bio-Chemistry Lab location has
  the org linked (prerequisites for billing/multi-user tests)
- Remove per-test beforeAll prereqs from billing tests (now in setup)
- Fix RESOURCE_CATEGORY_SLUG: use hyphen instead of underscore to
  match backend fixture slug (lab-tests-activity-definition)
